shared: move payroll events helpers to shared dir
