+++ future-apps/payroll/test/contracts/Payroll_allowed_tokens.test.js
@@ -1,5 +1,5 @@
 const PAYMENT_TYPES = require('../helpers/payment_types')
-const { getEvent } = require('../helpers/events')
+const { getEventAt } = require('../../../../shared/test-helpers/events')
 const { assertRevert } = require('@aragon/test-helpers/assertThrow')
 const { annualSalaryPerSecond } = require('../helpers/numbers')(web3)
 const { NOW, ONE_MONTH, RATE_EXPIRATION_TIME } = require('../helpers/time')
@@ -34,7 +34,7 @@ contract('Payroll allowed tokens,', ([owner, employee, anyone]) => {
           it('can allow a token', async () => {
             const receipt = await payroll.addAllowedToken(DAI.address, { from })
 
-            const event = getEvent(receipt, 'AddAllowedToken')
+            const event = getEventAt(receipt, 'AddAllowedToken')
             assert.equal(event.token, DAI.address, 'denomination token address should match')
 
             assert.equal(await payroll.getAllowedTokensArrayLength(), 1, 'allowed tokens length does not match')
@@ -44,7 +44,7 @@ contract('Payroll allowed tokens,', ([owner, employee, anyone]) => {
           it('can allow a the zero address', async () => {
             const receipt = await payroll.addAllowedToken(ZERO_ADDRESS, { from })
 
-            const event = getEvent(receipt, 'AddAllowedToken')
+            const event = getEventAt(receipt, 'AddAllowedToken')
             assert.equal(event.token, ZERO_ADDRESS, 'denomination token address should match')
 
             assert.equal(await payroll.getAllowedTokensArrayLength(), 1, 'allowed tokens length does not match')
