+++ future-apps/payroll/test/contracts/Payroll_rates.test.js
@@ -1,4 +1,4 @@
-const { getEvents } = require('../helpers/events')
+const { getEvents } = require('../../../../shared/test-helpers/events')
 const PAYMENT_TYPES = require('../helpers/payment_types')
 const { bigExp, ONE } = require('../helpers/numbers')(web3)
 const { NOW, TWO_MINUTES, RATE_EXPIRATION_TIME } = require('../helpers/time')
