+++ future-apps/payroll/test/contracts/Payroll_token_allocations.test.js
@@ -1,6 +1,6 @@
 const { assertRevert } = require('@aragon/test-helpers/assertThrow')
 const { annualSalaryPerSecond } = require('../helpers/numbers')(web3)
-const { getEvents, getEventArgument } = require('../helpers/events')
+const { getEvents, getEventArgument } = require('../../../../shared/test-helpers/events')
 const { NOW, ONE_MONTH, RATE_EXPIRATION_TIME } = require('../helpers/time')
 const { USD, deployDAI, deployTokenAndDeposit } = require('../helpers/tokens')(artifacts, web3)
 const { deployContracts, createPayrollAndPriceFeed } = require('../helpers/deploy')(artifacts, web3)
