--- future-apps/payroll/test/helpers/events.js
@@ -1,9 +1,0 @@
-const getEvent = (receipt, event) => getEvents(receipt, event)[0].args
-const getEvents = (receipt, event) => receipt.logs.filter(l => l.event === event)
-const getEventArgument = (receipt, event, arg) => getEvent(receipt, event)[arg]
-
-module.exports = {
-  getEvent,
-  getEvents,
-  getEventArgument,
-}
